refactor: Organize and document functions
