+++ Runtime/Math/Mesh/MeshUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bfbab31ecc3759a41ae4d24df5643bff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
